Move replacement logic to `MutableReplacementList`
